@@ -25,7 +25,7 @@
     "build:api-docs": "widdershins --code true --language_tabs --omitBody true --omitHeader true --resolve true --outfile docs/rest-api.md ui/api/openapi.json",
 
     "clean": "rm -rf .nuxt tmp",
-    "update-package-lock": "rm package-lock.json && npm install && npm audit fix && npx browserslist@latest --update-db",
+    "update-package-lock": "rm package-lock.json && npm install && npm audit fix && npx update-browserslist-db@latest",
 
     "lint": "npm run lint:eslint && npm run lint:stylelint",
     "lint:eslint": "eslint .",
@@ -43,57 +43,60 @@
     "@nuxtjs/axios": "^5.13.6",
     "@nuxtjs/robots": "^2.5.0",
     "@nuxtjs/sitemap": "^2.4.0",
-    "@octokit/rest": "~21.0.1",
-    "a11y-dialog": "^8.0.4",
-    "ajv": "^8.16.0",
+    "@octokit/rest": "~21.0.2",
+    "a11y-dialog": "^8.1.1",
+    "ajv": "^8.17.1",
     "ajv-formats": "^3.0.1",
     "chalk": "^5.3.0",
     "color-hash": "^2.0.2",
-    "color-name-list": "^10.24.0",
+    "color-name-list": "^10.28.1",
     "cookie-universal-nuxt": "^2.2.2",
     "cors": "^2.8.5",
     "embetty-vue": "^2.0.1",
-    "express": "^4.19.2",
+    "express": "^4.21.1",
     "jszip": "^3.10.1",
     "node-env-file": "^0.1.8",
     "nuxt": "~2.18.1",
     "nuxt-helmet": "^3.0.0",
-    "openapi-backend": "^5.10.6",
+    "openapi-backend": "^5.11.1",
     "sanitize-filename": "^1.6.3",
-    "sass": "^1.77.8",
+    "sass": "^1.81.0",
     "sass-loader": "^10.5.2",
     "scroll-into-view": "^1.16.2",
     "svg-inline-loader": "^0.8.2",
     "uuid": "^10.0.0",
     "vue-form": "^4.10.3",
-    "vue-ts-types": "^1.6.2",
+    "vue-ts-types": "^1.7.1",
     "vuedraggable": "^2.24.3",
     "xml2js": "^0.6.2",
     "xmlbuilder": "^15.1.1"
   },
   "devDependencies": {
-    "@stylistic/stylelint-config": "~1.0.1",
+    "@eslint/compat": "^1.2.3",
+    "@eslint/js": "~9.15.0",
+    "@eslint/markdown": "~6.2.1",
+    "@stylistic/stylelint-config": "~2.0.0",
     "broken-link-checker": "^0.7.8",
     "dir-compare": "^5.0.0",
     "disparity": "^3.2.0",
-    "eslint": "~8.57.0",
-    "eslint-plugin-import": "~2.29.1",
-    "eslint-plugin-jsdoc": "~48.0.2",
-    "eslint-plugin-jsonc": "~2.16.0",
-    "eslint-plugin-markdown": "~5.1.0",
+    "eslint": "~9.15.0",
+    "eslint-plugin-import": "~2.31.0",
+    "eslint-plugin-jsdoc": "~50.0.1",
+    "eslint-plugin-jsonc": "~2.18.2",
     "eslint-plugin-nuxt": "~4.0.0",
-    "eslint-plugin-promise": "~6.4.0",
-    "eslint-plugin-sonarjs": "~1.0.3",
-    "eslint-plugin-unicorn": "~54.0.0",
-    "eslint-plugin-vue": "~9.27.0",
-    "eslint-plugin-vuejs-accessibility": "~2.3.1",
-    "jsdoc-to-markdown": "^8.0.1",
+    "eslint-plugin-promise": "~7.1.0",
+    "eslint-plugin-sonarjs": "~1.0.4",
+    "eslint-plugin-unicorn": "~56.0.1",
+    "eslint-plugin-vue": "~9.31.0",
+    "eslint-plugin-vuejs-accessibility": "~2.4.1",
+    "globals": "^15.12.0",
+    "jsdoc-to-markdown": "^9.0.5",
     "libxmljs": "^1.0.11",
     "minimist": "^1.2.8",
     "postcss-html": "^1.7.0",
-    "stylelint": "~16.7.0",
+    "stylelint": "~16.10.0",
     "stylelint-config-html": "~1.1.0",
-    "stylelint-config-recess-order": "~5.0.1",
+    "stylelint-config-recess-order": "~5.1.1",
     "stylelint-config-standard-scss": "~13.1.0",
     "terminate": "^2.6.1",
     "widdershins": "^4.0.1"
