@@ -41,7 +41,10 @@ export async function fixtureFromRepository(manufacturerKey, fixtureKey) {
 
   if (fixtureJson.$schema.endsWith(`/fixture-redirect.json`)) {
     fixturePath = `../fixtures/${fixtureJson.redirectTo}.json`;
-    fixtureJson = Object.assign({}, await importJson(fixturePath, import.meta.url), { name: fixtureJson.name });
+    fixtureJson = {
+      ...await importJson(fixturePath, import.meta.url),
+      name: fixtureJson.name,
+    };
   }
 
   const manufacturer = await manufacturerFromRepository(manufacturerKey);
