@@ -387,7 +387,10 @@ class CoarseChannel extends AbstractChannel {
    */
   get capabilities() {
     if (`capability` in this._jsonObject) {
-      const capabilityData = Object.assign({ dmxRange: [0, Math.pow(256, this.dmxValueResolution) - 1] }, this._jsonObject.capability);
+      const capabilityData = {
+        dmxRange: [0, Math.pow(256, this.dmxValueResolution) - 1],
+        ...this._jsonObject.capability,
+      };
 
       return cacheResult(this, `capabilities`, [
         new Capability(capabilityData, this.dmxValueResolution, this),
