@@ -86,11 +86,14 @@ class Mode {
 
     const fixturePhysical = this.fixture.physical.jsonObject;
     const physicalOverride = this._jsonObject.physical;
-    const physicalData = Object.assign({}, fixturePhysical, physicalOverride);
+    const physicalData = { ...fixturePhysical, ...physicalOverride };
 
     for (const property of [`bulb`, `lens`, `matrixPixels`]) {
       if (property in physicalData) {
-        physicalData[property] = Object.assign({}, fixturePhysical[property], physicalOverride[property]);
+        physicalData[property] = {
+          ...fixturePhysical[property],
+          ...physicalOverride[property],
+        };
       }
     }
 
