@@ -177,6 +177,18 @@ function getCombinedEcueChannels(ecueFixture) {
 }
 
 /**
+ * @param {string | undefined} direction A string containing something like "CW", "CCW", "clockwise", "counter-clockwise".
+ * @returns {string} The normalized direction suffix.
+ */
+function getDirectionSuffix(direction) {
+  if (!direction) {
+    return ``;
+  }
+
+  return /^(?:clockwise|cw),?\s+$/i.test(direction) ? ` CW` : ` CCW`;
+}
+
+/**
  * Parses the e:cue channel and adds it to OFL fixture's availableChannels and the first mode.
  * @param {object} ecueChannel The e:cue channel object.
  * @param {object} fixture The OFL fixture object.
@@ -488,9 +500,8 @@ function addChannelToFixture(ecueChannel, fixture, warningsArray) {
      */
     function getSpeedGuessedComment() {
       return capabilityName.replace(/(?:^|,\s*|\s+)\(?((?:(?:counter-?)?clockwise|c?cw)(?:,\s*|\s+))?\(?(slow|fast|\d+|\d+\s*hz)\s*(?:-|to|–|…|\.{2,}|->|<->|→)\s*(fast|slow|\d+\s*hz)\)?$/i, (match, direction, start, end) => {
-        const directionString = direction ? (/^(?:clockwise|cw),?\s+$/i.test(direction) ? ` CW` : ` CCW`) : ``;
-
-        if (directionString !== ``) {
+        const directionSuffix = getDirectionSuffix(direction);
+        if (directionSuffix !== ``) {
           capability.type = `Rotation`;
         }
 
@@ -504,8 +515,8 @@ function addChannelToFixture(ecueChannel, fixture, warningsArray) {
           end = `${endNumber}Hz`;
         }
 
-        capability.speedStart = start + directionString;
-        capability.speedEnd = end + directionString;
+        capability.speedStart = start + directionSuffix;
+        capability.speedEnd = end + directionSuffix;
 
         // delete the parsed part
         return ``;
