@@ -3,9 +3,9 @@ import xml2js from 'xml2js';
 
 import importJson from '../../lib/import-json.js';
 import CoarseChannel from '../../lib/model/CoarseChannel.js';
-import { scaleDmxValue, scaleDmxRangeIndividually } from '../../lib/scale-dmx-values.js';
+import { scaleDmxRangeIndividually, scaleDmxValue } from '../../lib/scale-dmx-values.js';
 import gdtfAttributes, { gdtfUnits } from './gdtf-attributes.js';
-import { getRgbColorFromGdtfColor, followXmlNodeReference } from './gdtf-helpers.js';
+import { followXmlNodeReference, getRgbColorFromGdtfColor } from './gdtf-helpers.js';
 
 export const version = `0.2.0`;
 
@@ -670,11 +670,11 @@ export async function importFixtures(buffer, filename, authorName) {
         }
 
         // save the inherited result for later access
-        gdtfAttributes[attributeName] = Object.assign(
-          {},
-          getCapabilityTypeData(capabilityTypeData.inheritFrom),
-          capabilityTypeData,
-        );
+        gdtfAttributes[attributeName] = {
+
+          ...getCapabilityTypeData(capabilityTypeData.inheritFrom),
+          ...capabilityTypeData,
+        };
         delete gdtfAttributes[attributeName].inheritFrom;
 
         return gdtfAttributes[attributeName];
